# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Cat Shelter CRM",
        "contact": {
            "name": "Lev Yuvenskiy",
            "email": "levayu22@gmail.com"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:9000",
    "basePath": "/",
    "paths": {
        "/cats": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cats"
                ],
                "summary": "Get cats list",
                "operationId": "getCatsList",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Cat"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorJSON"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cats"
                ],
                "summary": "Create cat",
                "operationId": "createCat",
                "parameters": [
                    {
                        "description": "Parameters for creating a cat.",
                        "name": "createCatObject",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateCatRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.CatChipNumber"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorJSON"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorJSON"
                        }
                    }
                }
            }
        },
        "/cats/{chip_number}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cats"
                ],
                "summary": "Get cat",
                "operationId": "getCat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chip number",
                        "name": "chip_number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Cat"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorJSON"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cats"
                ],
                "summary": "Update cat",
                "operationId": "updateCat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chip number",
                        "name": "chip_number",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Parameters for updating a cat.",
                        "name": "createCatObject",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateCatRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorJSON"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cats"
                ],
                "summary": "Delete cat",
                "operationId": "deleteCat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chip number",
                        "name": "chip_number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorJSON"
                        }
                    }
                }
            }
        },
        "/guardians": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "guardians"
                ],
                "summary": "Get guardians list",
                "operationId": "getGuardiansList",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Guardian"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorJSON"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "guardians"
                ],
                "summary": "Create guardian",
                "operationId": "createGuardian",
                "parameters": [
                    {
                        "description": "Parameters for creating a guardian.",
                        "name": "createGuardianObject",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateGuardianRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.GuardianId"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorJSON"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorJSON"
                        }
                    }
                }
            }
        },
        "/guardians/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "guardians"
                ],
                "summary": "Get guardian",
                "operationId": "getGuardian",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Guardian"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorJSON"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "guardians"
                ],
                "summary": "Update guardian",
                "operationId": "updateGuardian",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Parameters for updating a guardian.",
                        "name": "createGuardianObject",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateGuardianRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorJSON"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "guardians"
                ],
                "summary": "Delete guardian",
                "operationId": "deleteGuardian",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorJSON"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apperror.ErrorJSON": {
            "type": "object",
            "properties": {
                "developerMessage": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.Cat": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "chipNumber": {
                    "type": "string"
                },
                "dateOfAdmissionToShelter": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "photoUrl": {
                    "type": "string"
                }
            }
        },
        "models.CatChipNumber": {
            "type": "object",
            "properties": {
                "chipNumber": {
                    "type": "string"
                }
            }
        },
        "models.CreateCatRequest": {
            "type": "object",
            "required": [
                "nickname"
            ],
            "properties": {
                "age": {
                    "type": "integer"
                },
                "chipNumber": {
                    "type": "string"
                },
                "dateOfAdmissionToShelter": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "photoUrl": {
                    "type": "string"
                }
            }
        },
        "models.CreateGuardianRequest": {
            "type": "object",
            "properties": {
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "photoUrl": {
                    "type": "string"
                }
            }
        },
...
        "models.Guardian": {
            "type": "object",
            "properties": {
                "firstname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastname": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "photoUrl": {
                    "type": "string"
                }
            }
        },
        "models.GuardianId": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
    }
}